LUAFDN-389 assign name to instance from their keys (#150)
The previous PR did not account for keys that comes from outside of a
component. With these changes, name should have the first key coming
from their parent in the component hierarchy, just like legacy Roact
does it.

This also change a bit how `RoactCompat.mount` works so that it is more 
similar to legacy Roact.