Move Bindings, adjust type logic, add fixmes (#156)
Closes LUAFDN-390.

This moves the logic for bindings into the React project and re-implements refs on top of them. With this logic in place, values that are assigned as refs will also update dynamically when the reconciler subscribes to their value. I've added a couple of tests to demonstrate what sorts of use cases this feature addresses.