Rotriever alpha 5 (#158)
Upgrades to the newest Rotriever, which restores type exporting and works around behaviors in roblox-cli that were causing some files to not be typechecked or linted.

Rotriever used to symlink path dependencies directly to their declared content_root. Now, we link to the root of the project and trust that the project files there will nest correctly.

It appears that if roblox-cli analyze follows nested projects via symlinks, it loses track of the .robloxrc config and does not type check. This is why there are new lints; I'm not sure if this is the result of symlinks, nested projects, or some combination thereof. To work around it, I've added .robloxrc files to each of the individual projects.

This surfaced a number of new lints, most of them valid. I've cleaned up the legitimate ones, and added workarounds for the rest.