LUAFDN-395: Prevent components mistakenly using reserved field names (#154)
Ticket: https://jira.rbx.com/browse/LUAFDN-395

updater is a reserved name that is internal only so to discourage external usage and avoid potential conflicts, we've changed that to __updater.

refs is a reserved name, as it is used for legacy stringrefs. However, some of our components set self.refs to something of their own, causing breaks and failures.

To resolve this, we simply remove refs. It is a legacy feature that adds no additional functionality, so removing it does not have any impact on our codebase's capabilities or compatibility. We didn't use it anyway, and it isn't worth supporting if it brings up problems for no real gain.