LUAFDN-498 - Better error output for bad prop assignments (#200)
Improves the error output when assigning a bad prop. Crucially, this also replaces the hard error with a console.error instead.

Before these changes, the UI would fail to render and have no meaningful way to recover unless an error boundary has been established. With these changes, we print an accordingly but swallow the actual hard error. This means that we're a lot more likely to get mostly-working behavior, which may be preferable for production environments.

We've kept the hard error for now to avoid losing signal from production / canary testing, but we can remove it once we introduce that logic.