Remove troublesome optimization (#221)
Reverts a minor optimization to bindings that prevented them from propagating updates when the assigned value was the same; this can cause code that relies on other updates to the same fields (e.g. the Text field of a TextBox) to fail to update in expected ways.

The best approach is to roll back the optimization and revisit it later when we're thinking more about controlled components. Here's my reasoning:

Bindings work pretty much identically to legacy Roact (once we roll this back), so we should think through how/when we make behavior changes in the name of optimization
This problem intersects with the controlled/uncontrolled problem pretty directly, and should be addressed alongside it
Bindings are already a pretty-fast escape hatch, and they're used pretty sparingly. I don't have a sense that the added optimization is a good use of our energy with perf fixes.