#nojira - Add global for legacy root (#233)
Adds a global to toggle whether or not we return a legacy root. In practice, this will be used during the upgrade instead of __ROACT_17_INLINE_ACT__. The goal as we fix concurrent tests will be to migrate all test suites from __ROACT_17_COMPAT_LEGACY_ROOT__ to __ROACT_17_INLINE_ACT__. This flag lets us pin behavior along the way.