LUAFDN-715 Fix Mysterious top bar disappearing bug (#249)
Problem:

In certain games, the top bar was disappearing.

https://jira.rbx.com/browse/LUAFDN-715

After a lot of debugging, I eventually found that this was caused by the EmotesMenu unmounting the top bar. It turned out that the top bar and the emotes menu both had the same root!

Solution:
The cause of this is that it is setmetatable(table, metatable), not setmetatable(metatable, table). Tables that wish to share behaviour but not be the same table need to share a metatable, not be the same table!