LUAFDN-692: Add assertions around the instance map to detect bad objects being inserted (#248)
This doesn't fix the production crash we saw in LUAFDN-692, but it will give us more information to track down the root cause.

We get this crash in production:
https://console.crashes.rbxinfra.com/p/UniversalApp/debug?fingerprint=a7b1edb8ef4d68db06604a4ec8bc8c16117e5b7f60b91f441ee4de889e15f91c&debug=(%22288330f%22,0,3)

Tracing through that, an instance comes out of the map whose alternate's childLanes is nil. childLanes is initialized by default on every Fiber, and is only assigned to from other fiber.childLanes or from bit32.* calls. Lua Apps doesn't directly contain code that matches alternate = or \[.*alternate.*\].*=. This tells me that the alternate is probably not a fiber, but some other kind of table. I think the best bet is to catch when an invalid fiber or invalid fiber.alternate gets puts into the map.

I suspect that we'll need to see what the inner component of the RoactRodux Connect component is to fully trace it into Lua Apps. As such, I've put up a draft PR to roact-rodux to try harder to get the inner component name and embed it in the Connect component name:
Roblox/roact-rodux#59

If this proves successful in tracking down the bug, we can keep this check in but only in DEV, and submit a PR upstream.