Separate scheduler mocking from act inlining (#259)
This helps us address some issues we encountered in the wild (https://jira.rbx.com/browse/SOCRTC-1885) that suggest that we might not want to inline act in all cases.

Ideally we'd always mock the scheduler when we're in a testing context, but that hits some snags with some of our existing tests and is less trivial to address. For now, I want to split up the behavior so it's easier to opt into the parts of it that are needed.