LUAFDN-408 - useBinding hook (#304)
Introduces a useBinding hook. I wanted to see how bad this would be... turns out: not very!

I added a bunch of type strictness to bindings themselves.

They're currently constructed very defensively, with indirection onto a symbol key to make it harder to access internals.
Because of this, I needed to do some pretty "just trust me" casting, but I think it makes for a very reasonable public type interface
There are a few places where we lost specificity because the map function can't be defined the way I thought it should be
The hyper-defensive structure of bindings could probably be relaxed for slightly improved perf and better internal ergonomics, curious what folks think about that
Also worth noting: with the useBinding hook, you'll always get the same binding you created in the first render. However, the map method will still always create new bindings. Just like in legacy Roact, this is considered acceptable; since bindings are designed to allow short-circuiting of render itself, the binding being redefined on render (when non-binding changes required a re-render) should not negate any of the benefits of using it in the first place.

Checklist before submitting:

 Added/updated relevant tests
 Added/updated documentation