APPFDN-3248 React Profiler flags (#426)
This PR makes it possible to use `React.Profiler` in production code by
setting the `_G.__PROFILE__` flag. In particular, we:
1. Disable `enableProfiling` from SchedulerFeatureFlags.lua to be
consistent with
[upstream](https://github.com/facebook/react/blob/main/packages/scheduler/src/SchedulerFeatureFlags.js#L10)
3. Disable `enableSchedulerTracing` since it was removed in newer
versions [upstream](https://github.com/facebook/react/pull/20037)
2. Enable `enableProfilerCommitHooks` to be consistent with
[upstream](https://github.com/facebook/react/blob/main/packages/shared/ReactFeatureFlags.js#L252-L253)

I also deleted an old workflow which was failing OSS leak scan. It was
set up to send PR review reminders to a now-defunct channel, so it was
obviously not being used. We can add this back later if the team decides
they want it in the future.

Checklist before submitting:
* [ ] Added/updated relevant tests
* [ ] Added/updated documentation