Fix useCallback errors in DEV mode (#474)
### Why are the changes being made?

When re-rendering a component using derived state idiom for functional
components, if the component had any useCallbacks inside it, React would
error with `Rendered more hooks than during the previous render.` when
in DEV mode.

It looks like this was caused by a mistranslation where instead of
calling updateCallback, we would call mountCallback in DEV mode. Note
that in non-DEV mode, we already call the correct api (updateCallback)
and this seems to be a mistranslation since this was implemented
correctly in upstream:
https://github.com/facebook/react/blob/43363e2795393a00fd77312a16d6b80e626c29de/packages/react-reconciler/src/ReactFiberHooks.new.js#L2289

Unflagged because it's under DEV mode.